Css login

+++ model/creation.sql
@@ -36,14 +36,14 @@ CREATE TABLE POINT_DE_COLLECTE (
   nom_pt_collecte VARCHAR(42) UNIQUE,
   pos_x DECIMAL(10,4),
   pos_y DECIMAL(10,4),
-  qte_max INT
+  qte_max DECIMAL(10,4)
 );
 
 CREATE TABLE DECHET (
   id_Dechet  INT AUTO_INCREMENT PRIMARY KEY,
   nom_Dechet VARCHAR(42),
   id_Type INT NOT NULL,
-  qte INT,
+  qte DECIMAL(10,4),
   FOREIGN KEY (id_Type) REFERENCES CATEGORIEDECHET (id_Type)
 );
 
@@ -60,7 +60,7 @@ CREATE TABLE DEPOSER (
 CREATE TABLE TRAITER (
   id_point_collecte INT NOT NULL,
   id_Type       INT NOT NULL,
-  qte_collecte   INT,
+  qte_collecte   DECIMAL(10,4),
   date_collecte DATE,
   PRIMARY KEY (id_point_collecte, id_Type),
   FOREIGN KEY (id_point_collecte) REFERENCES POINT_DE_COLLECTE (id_point_collecte),
